+++ packages/cli/core/src/commands/org/resources/preview.spec.ts
@@ -26,7 +26,7 @@ import {CliUx, Command} from '@oclif/core';
 import {AuthenticatedClient} from '@coveo/cli-commons/platform/authenticatedClient';
 import {IsGitInstalled} from '../../../lib/decorators/preconditions';
 
-const mockedSnapshotFactory = jest.mocked(SnapshotFactory, true);
+const mockedSnapshotFactory = jest.mocked(SnapshotFactory);
 const mockedConfig = jest.mocked(Config);
 const mockedProject = jest.mocked(Project);
 const mockedConfigGet = jest.fn();
@@ -38,7 +38,7 @@ const mockedAreResourcesInError = jest.fn();
 const mockedValidateSnapshot = jest.fn();
 const mockedPreviewSnapshot = jest.fn();
 const mockedLastReport = jest.fn();
-const mockedIsGitInstalled = jest.mocked(IsGitInstalled, true);
+const mockedIsGitInstalled = jest.mocked(IsGitInstalled);
 const mockedAuthenticatedClient = jest.mocked(AuthenticatedClient);
 const mockEvaluate = jest.fn();
 
@@ -87,7 +87,7 @@ const mockUserNotHavingAllRequiredPlatformPrivileges = () => {
   mockEvaluate.mockResolvedValue({approved: false});
 };
 
-const mockSnapshotFactory = async () => {
+const mockSnapshotFactory = () => {
   const dummySnapshot = {
     validate: mockedValidateSnapshot,
     preview: mockedPreviewSnapshot,
@@ -106,7 +106,7 @@ const mockSnapshotFactory = async () => {
   );
 };
 
-const mockSnapshotFactoryReturningValidSnapshot = async () => {
+const mockSnapshotFactoryReturningValidSnapshot = () => {
   const successReport = getSuccessReport(
     'success-report',
     ResourceSnapshotsReportType.Apply
@@ -114,10 +114,10 @@ const mockSnapshotFactoryReturningValidSnapshot = async () => {
   mockedValidateSnapshot.mockImplementation(() =>
     Promise.resolve(new SnapshotReporter(successReport))
   );
-  await mockSnapshotFactory();
+  mockSnapshotFactory();
 };
 
-const mockSnapshotFactoryReturningInvalidSnapshot = async () => {
+const mockSnapshotFactoryReturningInvalidSnapshot = () => {
   const errorReport = getErrorReport(
     'error-report',
     ResourceSnapshotsReportType.Apply
@@ -125,20 +125,19 @@ const mockSnapshotFactoryReturningInvalidSnapshot = async () => {
   mockedValidateSnapshot.mockImplementation(() =>
     Promise.resolve(new SnapshotReporter(errorReport))
   );
-  await mockSnapshotFactory();
+  mockSnapshotFactory();
 };
 
-const mockSnapshotFactoryReturningSnapshotWithMissingVaultEntries =
-  async () => {
-    const missingVaultEntry = getMissingVaultEntryReport(
-      'missing-vault-entry',
-      ResourceSnapshotsReportType.Apply
-    );
-    mockedValidateSnapshot.mockImplementation(() =>
-      Promise.resolve(new SnapshotReporter(missingVaultEntry))
-    );
-    await mockSnapshotFactory();
-  };
+const mockSnapshotFactoryReturningSnapshotWithMissingVaultEntries = () => {
+  const missingVaultEntry = getMissingVaultEntryReport(
+    'missing-vault-entry',
+    ResourceSnapshotsReportType.Apply
+  );
+  mockedValidateSnapshot.mockImplementation(() =>
+    Promise.resolve(new SnapshotReporter(missingVaultEntry))
+  );
+  mockSnapshotFactory();
+};
 
 describe('org:resources:preview', () => {
   const doMockPreconditions = function () {
@@ -163,8 +162,8 @@ describe('org:resources:preview', () => {
   });
 
   describe('when the report contains no resources in error', () => {
-    beforeAll(async () => {
-      await mockSnapshotFactoryReturningValidSnapshot();
+    beforeAll(() => {
+      mockSnapshotFactoryReturningValidSnapshot();
     });
 
     afterAll(() => {
@@ -339,8 +338,8 @@ describe('org:resources:preview', () => {
   });
   //#region TODO: CDX-948, setup phase needs to be rewrite and assertions 'split up' (e.g. the error ain't trigger directly by the function, therefore should not be handled)
   describe('when the report contains resources in error', () => {
-    beforeAll(async () => {
-      await mockSnapshotFactoryReturningInvalidSnapshot();
+    beforeAll(() => {
+      mockSnapshotFactoryReturningInvalidSnapshot();
     });
 
     beforeEach(() => {
@@ -374,8 +373,8 @@ describe('org:resources:preview', () => {
   });
 
   describe('when the report contains resources with missing vault entries', () => {
-    beforeAll(async () => {
-      await mockSnapshotFactoryReturningSnapshotWithMissingVaultEntries();
+    beforeAll(() => {
+      mockSnapshotFactoryReturningSnapshotWithMissingVaultEntries();
     });
 
     afterAll(() => {
@@ -386,7 +385,7 @@ describe('org:resources:preview', () => {
       test
         .stdout()
         .stderr()
-        .stub(CliUx.ux, 'confirm', () => async () => false)
+        .stub(CliUx.ux, 'confirm', () => () => Promise.resolve(false))
         .command(['org:resources:preview'])
         .catch(/Your destination organization is missing vault entries/)
         .it('should throw an error for invalid snapshots');