+++ packages/cli/core/src/commands/org/resources/pull.ts
@@ -1,5 +1,6 @@
-import {ResourceSnapshotType} from '@coveord/platform-client';
-import {Flags, CliUx} from '@oclif/core';
+import type {ResourceSnapshotType} from '@coveord/platform-client';
+import {startSpinner, stopSpinner} from '@coveo/cli-commons/utils/ux';
+import {Flags} from '@oclif/core';
 import {blueBright} from 'chalk';
 import {readJsonSync} from 'fs-extra';
 import {resolve} from 'path';
@@ -84,7 +85,7 @@ export default class Pull extends CLICommand {
       default: allowedResourceType,
     }),
     model: Flags.custom<SnapshotPullModel>({
-      parse: async (input: string): Promise<SnapshotPullModel> => {
+      parse: (input: string): Promise<SnapshotPullModel> => {
         const model = readJsonSync(resolve(input));
         validateSnapshotPullModel(model);
         return model;
@@ -111,15 +112,14 @@ export default class Pull extends CLICommand {
 
     const snapshot = await this.getSnapshot();
 
-    CliUx.ux.action.start('Updating project with Snapshot');
+    startSpinner('Updating project with Snapshot');
     await this.refreshProject(project, snapshot);
     project.writeResourcesManifest(targetOrganization);
 
     if (await this.shouldDeleteSnapshot()) {
       await snapshot.delete();
     }
-
-    CliUx.ux.action.stop('Project updated');
+    stopSpinner({message: 'Project updated'});
   }
 
   private async shouldDeleteSnapshot() {
@@ -171,7 +171,7 @@ export default class Pull extends CLICommand {
     );
   }
 
-  private async askUserShouldWeOverwrite() {
+  private askUserShouldWeOverwrite() {
     const question = PullCommandStrings.projectOverwriteQuestion(
       Project.resourceFolderName
     );
@@ -188,13 +188,13 @@ export default class Pull extends CLICommand {
 
   private async createAndGetNewSnapshot(target: string) {
     const resourcesToExport = await this.getResourceSnapshotTypesToExport();
-    CliUx.ux.action.start(`Creating Snapshot from ${formatOrgId(target)}`);
+    startSpinner(`Creating Snapshot from ${formatOrgId(target)}`);
     const waitOption = await this.getWaitOption();
     return SnapshotFactory.createFromOrg(resourcesToExport, target, waitOption);
   }
 
   private async getExistingSnapshot(snapshotId: string, target: string) {
-    CliUx.ux.action.start('Retrieving Snapshot');
+    startSpinner('Retrieving Snapshot');
     const waitOption = await this.getWaitOption();
     return SnapshotFactory.createFromExistingSnapshot(
       snapshotId,