+++ packages/cli/core/src/lib/oauth/oauthClientServer.spec.ts
@@ -23,7 +23,7 @@ import {AuthorizationServiceConfiguration, ClientConfig} from './oauthConfig';
 type createServerInitialOverload = jest.MaybeMocked<{
   (requestListener?: RequestListener | undefined): Server;
 }>;
-const mockedAxios = jest.mocked(axios, true);
+const mockedAxios = jest.mocked(axios);
 const mockedCreateServer = jest.mocked(
   createServer
 ) as unknown as createServerInitialOverload;