+++ packages/cli/source/jest.config.js
@@ -6,11 +6,6 @@ process.env.TS_NODE_PROJECT = require.resolve('./tsconfig.dev.json');
 module.exports = {
   rootDir: '.',
   preset: 'ts-jest',
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.dev.json',
-    },
-  },
   testEnvironment: 'node',
   verbose: true,
   collectCoverage: true,